Add background traffic for test_pfc_pause_extra_lossless test
Add background traffic for test_pfc_pause_extra_lossless test in
script tests/qos/test_tunnel_qos_remap.py
